+++ include/rl/logging/client/ema.h
@@ -17,36 +17,36 @@ namespace rl::logging::client
     class EMA : public Base
     {
         public:
-            EMA(const std::vector<float> &smoothing_values, int output_period_s);
+            EMA(const std::vector<double> &smoothing_values, int output_period_s);
             ~EMA();
             
-            void log_scalar(const std::string &name, float value) override;
+            void log_scalar(const std::string &name, double value) override;
 
             class Estimator {
                 public:
-                    Estimator(float smoothing) : smoothing{smoothing}, _x{smoothing}, inv_smoothing{1 - smoothing} {}
+                    Estimator(double smoothing) : smoothing{smoothing}, _x{smoothing}, inv_smoothing{1 - smoothing} {}
 
-                    inline float get() const { return value / (1 - _x); }
-                    inline void update(float observation) {
+                    inline double get() const { return value / (1 - _x); }
+                    inline void update(double observation) {
                         value += inv_smoothing * (observation - value);
                         _x *= smoothing;
                     }
 
                 private:
-                    const float inv_smoothing;
-                    const float smoothing;
-                    float value{0.0};
-                    float _x;
+                    const double inv_smoothing;
+                    const double smoothing;
+                    double value{0.0};
+                    double _x;
             };
 
         private:
             std::mutex estimate_update_mtx{};
             std::map<std::string, std::vector<Estimator>> estimates;
             
-            std::vector<float> smoothing_values;
+            std::vector<double> smoothing_values;
             int output_period;
             
-            thread_safe::Queue<std::pair<std::string, float>> scalar_queue{1000};
+            thread_safe::Queue<std::pair<std::string, double>> scalar_queue{1000};
             std::thread queue_consuming_thread;
             std::thread output_producing_thread;
